CDC #113 - Extracting MapControl component to allow for custom controls
